+++ src/GetDocumentInsider/Resources.resx
@@ -120,9 +120,6 @@
   <data name="AssemblyDescription" xml:space="preserve">
     <value>The assembly to use.</value>
   </data>
-  <data name="JsonDescription" xml:space="preserve">
-    <value>Show JSON output.</value>
-  </data>
   <data name="MissingOption" xml:space="preserve">
     <value>Missing required option '--{0}'.</value>
   </data>
@@ -135,38 +132,17 @@
   <data name="PrefixDescription" xml:space="preserve">
     <value>Prefix console output with logging level.</value>
   </data>
-  <data name="UsingApplicationBase" xml:space="preserve">
-    <value>Using application base '{0}'.</value>
-  </data>
-  <data name="UsingAssembly" xml:space="preserve">
-    <value>Using assembly '{0}'.</value>
-  </data>
-  <data name="UsingConfigurationFile" xml:space="preserve">
-    <value>Using configuration file '{0}'.</value>
-  </data>
   <data name="VerboseDescription" xml:space="preserve">
     <value>Show verbose output.</value>
   </data>
-  <data name="WritingFile" xml:space="preserve">
-    <value>Writing '{0}'...</value>
-  </data>
-  <data name="UsingWorkingDirectory" xml:space="preserve">
-    <value>Using working directory '{0}'.</value>
-  </data>
   <data name="ToolsDirectoryDescription" xml:space="preserve">
     <value>Location from which inside man was copied (in the .NET Framework case) or loaded.</value>
-  </data>
-  <data name="UriDescription" xml:space="preserve">
-    <value>The URI to download the document from.</value>
   </data>
   <data name="MethodDescription" xml:space="preserve">
     <value>The name of the method to invoke on the '--service' instance. Default value '{0}'.</value>
   </data>
   <data name="ServiceDescription" xml:space="preserve">
     <value>The qualified name of the service type to retrieve from dependency injection. Default value '{0}'.</value>
-  </data>
-  <data name="MissingOptions" xml:space="preserve">
-    <value>Missing required option '--{0}' or '--{1}'.</value>
   </data>
   <data name="DocumentDescription" xml:space="preserve">
     <value>The name of the document to pass to the '--method' method. Default value '{0}'.</value>
@@ -179,9 +155,6 @@
   </data>
   <data name="UsingService" xml:space="preserve">
     <value>Using service '{0}'.</value>
-  </data>
-  <data name="UsingUri" xml:space="preserve">
-    <value>Using URI '{0}'.</value>
   </data>
   <data name="MethodInvocationFailed" xml:space="preserve">
     <value>Method '{0}' of service '{1}' failed to generate document '{2}'.</value>
