--- src/GetDocumentInsider/WrappedException.cs
@@ -1,24 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-
-namespace Microsoft.Extensions.ApiDescription.Client
-{
-    internal class WrappedException : Exception
-    {
-        private readonly string _stackTrace;
-
-        public WrappedException(string type, string message, string stackTrace)
-            : base(message)
-        {
-            Type = type;
-            _stackTrace = stackTrace;
-        }
-
-        public string Type { get; }
-
-        public override string ToString()
-            => _stackTrace;
-    }
-}
