+++ src/Microsoft.Extensions.ApiDescription.Client/DownloadFile.cs
@@ -4,6 +4,9 @@
 using System;
 using System.IO;
 using System.Net.Http;
+using System.Net.Sockets;
+using System.Reflection;
+using System.Security.Cryptography;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Build.Framework;
@@ -94,19 +97,130 @@ namespace Microsoft.Extensions.ApiDescription.Client
                 return;
             }
 
-            log.LogMessage($"Downloading '{uri}' to '{destinationPath}'.");
+            log.LogMessage(MessageImportance.High, $"Downloading '{uri}' to '{destinationPath}'.");
 
-            using (var httpClient = new HttpClient
+            using (var httpClient = new HttpClient())
             {
-            })
+                await DownloadAsync(uri, destinationPath, httpClient, cancellationToken, log, timeoutSeconds);
+            }
+        }
+
+        public static async Task DownloadAsync(
+            string uri,
+            string destinationPath,
+            HttpClient httpClient,
+            CancellationToken cancellationToken,
+            TaskLoggingHelper log,
+            int timeoutSeconds)
+        {
+            // Timeout if the response has not begun within 1 minute
+            httpClient.Timeout = TimeSpan.FromMinutes(1);
+
+            var destinationExists = File.Exists(destinationPath);
+            var reachedCopy = false;
+            try
             {
-                await DownloadFileCore.DownloadAsync(
-                    uri,
-                    destinationPath,
-                    httpClient,
-                    new LogWrapper(log),
-                    cancellationToken,
-                    timeoutSeconds);
+                using (var response = await httpClient.GetAsync(uri, cancellationToken))
+                {
+                    response.EnsureSuccessStatusCode();
+                    cancellationToken.ThrowIfCancellationRequested();
+
+                    using (var responseStreamTask = response.Content.ReadAsStreamAsync())
+                    {
+                        var finished = await Task.WhenAny(
+                            responseStreamTask,
+                            Task.Delay(TimeSpan.FromSeconds(timeoutSeconds)));
+
+                        if (!ReferenceEquals(responseStreamTask, finished))
+                        {
+                            throw new TimeoutException($"Download failed to complete in {timeoutSeconds} seconds.");
+                        }
+
+                        using (var responseStream = await responseStreamTask)
+                        {
+                            if (destinationExists)
+                            {
+                                // Check hashes before using the downloaded information.
+                                var downloadHash = GetHash(responseStream);
+                                responseStream.Position = 0L;
+
+                                byte[] destinationHash;
+                                using (var destinationStream = File.OpenRead(destinationPath))
+                                {
+                                    destinationHash = GetHash(destinationStream);
+                                }
+
+                                var sameHashes = downloadHash.Length == destinationHash.Length;
+                                for (var i = 0; sameHashes && i < downloadHash.Length; i++)
+                                {
+                                    sameHashes = downloadHash[i] == destinationHash[i];
+                                }
+
+                                if (sameHashes)
+                                {
+                                    log.LogMessage($"Not overwriting existing and matching file '{destinationPath}'.");
+                                    return;
+                                }
+                            }
+                            else
+                            {
+                                // May need to create directory to hold the file.
+                                var destinationDirectory = Path.GetDirectoryName(destinationPath);
+                                if (!string.IsNullOrEmpty(destinationDirectory))
+                                {
+                                    Directory.CreateDirectory(destinationDirectory);
+                                }
+                            }
+
+                            // Create or overwrite the destination file.
+                            reachedCopy = true;
+                            using (var outStream = File.Create(destinationPath))
+                            {
+                                await responseStream.CopyToAsync(outStream);
+                            }
+                        }
+                    }
+                }
+            }
+            catch (HttpRequestException ex) when (destinationExists)
+            {
+                if (ex.InnerException is SocketException socketException)
+                {
+                    log.LogWarning($"Unable to download {uri}, socket error code '{socketException.SocketErrorCode}'.");
+                }
+                else
+                {
+                    log.LogWarning($"Unable to download {uri}: {ex.Message}");
+                }
+            }
+            catch (Exception ex)
+            {
+                log.LogError($"Downloading '{uri}' failed.");
+                log.LogErrorFromException(ex, showStackTrace: true);
+                if (reachedCopy)
+                {
+                    File.Delete(destinationPath);
+                }
+            }
+        }
+
+        private static byte[] GetHash(Stream stream)
+        {
+            SHA256 algorithm;
+            try
+            {
+                algorithm = SHA256.Create();
+            }
+            catch (TargetInvocationException)
+            {
+                // SHA256.Create is documented to throw this exception on FIPS-compliant machines. See
+                // https://msdn.microsoft.com/en-us/library/z08hz7ad Fall back to a FIPS-compliant SHA256 algorithm.
+                algorithm = new SHA256CryptoServiceProvider();
+            }
+
+            using (algorithm)
+            {
+                return algorithm.ComputeHash(stream);
             }
         }
     }
