--- src/dotnet-getdocument/ServiceProjectReferenceMetadata.props
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <!-- No need for any additional work in outer build. -->
-    <WriteServiceProjectReferenceMetadataDependsOn Condition="'$(TargetFramework)' != '' AND '$(WriteServiceProjectReferenceMetadataDependsOn)' == ''">
-      $(WriteServiceProjectReferenceMetadataDependsOn)
-    </WriteServiceProjectReferenceMetadataDependsOn>
-
-    <!-- Execute WriteServiceProjectReferenceMetadata support targets after WriteServiceProjectReferenceMetadataDependsOn targets by default. -->
-    <GetServiceProjectReferenceMetadataDependsOn Condition="'$(GetServiceProjectReferenceMetadataDependsOn)' == ''">
-      $(WriteServiceProjectReferenceMetadataDependsOn)
-    </GetServiceProjectReferenceMetadataDependsOn>
-    <WriteServiceProjectReferenceMetadataCoreDependsOn Condition="'$(WriteServiceProjectReferenceMetadataCoreDependsOn)' == ''">
-      $(WriteServiceProjectReferenceMetadataDependsOn)
-    </WriteServiceProjectReferenceMetadataCoreDependsOn>
-  </PropertyGroup>
-</Project>
